+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v 2.9.0 (?)
+# v 2.9.0 (2023-11-29)
 Changes in this release:
 - Add support for more values (yes, 1) and (no, 0) for boolean options
 - Add `deploy_resouce_v2` endpoint